test perf on long urls
